tests : add support for qwen3 SSM archs (#24031)
* tests : add support for qwen3 SSM archs

* arch : add LLM_KV_ATTENTION_RECURRENT_LAYERS

* cont : naming + TODOs

(cherry picked from commit 06938ac129e5feee1e731323e5c37dc973de5573)

+++ src/llama-model.cpp
@@ -2069,18 +2069,18 @@ llama_memory_i * llama_model::create_memory(const llama_memory_params & params, 
                         filter_attn = [&](uint32_t) { return true; };
                         filter_recr = [&](uint32_t) { return true; };
                     } else if (arch == LLM_ARCH_NEMOTRON_H || arch == LLM_ARCH_NEMOTRON_H_MOE) {
-                        filter_attn = [&](uint32_t il) {
+                        filter_attn = [&](int32_t il) {
                             return !hparams.is_recr(il) && hparams.n_ff(il) == 0;
                         };
-                        filter_recr = [&](uint32_t il) {
+                        filter_recr = [&](int32_t il) {
                             return hparams.is_recr(il) && hparams.n_ff(il) == 0;
                         };
                     } else if (arch == LLM_ARCH_QWEN35 || arch == LLM_ARCH_QWEN35MOE) {
-                        filter_attn = [&](uint32_t il) {
-                            return il < hparams.n_layer() && !hparams.is_recr(il);
+                        filter_attn = [&](int32_t il) {
+                            return (uint32_t)il < hparams.n_layer() && !hparams.is_recr(il);
                         };
-                        filter_recr = [&](uint32_t il) {
-                            return il < hparams.n_layer() && hparams.is_recr(il);
+                        filter_recr = [&](int32_t il) {
+                            return (uint32_t)il < hparams.n_layer() && hparams.is_recr(il);
                         };
                     }
